backend/requirements.txt
fastapi
uvicorn[standard]
supabase
openai
pydantic
pydantic-settings
python-dotenv
alembic
sqlalchemy
google-genai
httpx
python-multipart
pytest
pytest-asyncio
